# 09._OpenApi/python/openapi.json
{
  "openapi": "3.0.2",
  "info": { "title": "FastAPI", "version": "0.1.0" },
  "paths": {
    "/api/spacecrafts": {
      "get": {
        "tags": ["Spacecrafts"],
        "summary": "Get Spacecrafts",
        "operationId": "get_spacecrafts_api_spacecrafts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Spacecrafts Api Spacecrafts Get",
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Spacecraft" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Spacecrafts"],
        "summary": "Add Spacecraft",
        "operationId": "add_spacecraft_api_spacecrafts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Spacecraft" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Spacecraft" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/spacecrafts/{spacecraft_id}": {
      "get": {
        "tags": ["Spacecrafts"],
        "summary": "Get Spacecraft",
        "operationId": "get_spacecraft_api_spacecrafts__spacecraft_id__get",
        "parameters": [
          {
            "required": true,
            "schema": { "title": "Spacecraft Id", "type": "integer" },
            "name": "spacecraft_id",
            "in": "path"
          },
          {
            "description": "Show spacecraft id",
            "required": false,
            "schema": {
              "title": "Show Id",
              "type": "string",
              "description": "Show spacecraft id",
              "default": "Default"
            },
            "name": "show_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Spacecraft" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": { "$ref": "#/components/schemas/ValidationError" }
          }
        }
      },
      "Spacecraft": {
        "title": "Spacecraft",
        "required": ["id", "name"],
        "type": "object",
        "properties": {
          "id": { "title": "Id", "type": "integer" },
          "name": { "title": "Name", "type": "string" }
        }
      },
      "ValidationError": {
        "title": "ValidationError",
        "required": ["loc", "msg", "type"],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] }
          },
          "msg": { "title": "Message", "type": "string" },
          "type": { "title": "Error Type", "type": "string" }
        }
      }
    }
  }
}